+++ tests/acceptance/test_preprocessing.py
@@ -18,11 +18,10 @@ logger = getLogger("Logprep-Test")
 def get_config():
     pipeline = [
         {
-            "normalizername": {
-                "type": "normalizer",
-                "specific_rules": ["tests/testdata/acceptance/normalizer/rules_static/specific"],
-                "generic_rules": ["tests/testdata/acceptance/normalizer/rules_static/generic"],
-                "regex_mapping": "tests/testdata/acceptance/normalizer/rules_static/regex_mapping.yml",
+            "dissector": {
+                "type": "dissector",
+                "specific_rules": ["tests/testdata/acceptance/dissector/rules/specific"],
+                "generic_rules": ["tests/testdata/acceptance/dissector/rules/generic"],
             }
         }
     ]